requirements.txt
cffi==1.17.1
cryptography==45.0.6
hexdump==3.3
pycparser==2.22
pycryptodome==3.23.0
